requirements.txt
openai
python-dotenv
colorama
fastapi
uvicorn
python-multipart
gTTS
pydantic
